requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asgi-lifespan==2.1.0
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
gotrue==2.11.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.24
langchain-openai==0.2.12
langgraph==0.2.58
langgraph-checkpoint==2.0.8
langgraph-sdk==0.1.43
langsmith==0.2.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.0.2
openai==1.57.2
orjson==3.10.12
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pluggy==1.5.0
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
realtime==2.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==13.1
yarl==1.18.3
